Add config pages to main menu UI for initial set-up
